# internal/server/handlers/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис хранения метрик.",
        "title": "MetricStorage API",
        "contact": {
            "email": "s.turchinskiy@yandex.ru"
        },
        "version": "1.0"
    },
    "host": "nohost.io:8080",
    "basePath": "/",
    "paths": {
...
        "/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создание новой / обновление существующей метрики из json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Сохранение метрики",
                "operationId": "updateUpdateMetricJSON",
                "parameters": [
                    {
                        "description": "Содержимое метрики",
                        "name": "metric_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Metrics"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Metrics"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/updates": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Массовое создание / обновление существующих метрик из json",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Сохранение метрик",
                "operationId": "updateUpdateMetricsBatch",
                "parameters": [
                    {
                        "description": "Метрики",
                        "name": "metric_data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Metrics"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Load 5 records",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Ошибка авторизации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Metrics": {
            "type": "object",
            "properties": {
                "delta": {
                    "description": "значение метрики в случае передачи counter",
                    "type": "integer",
                    "example": 100
                },
                "id": {
                    "description": "имя метрики",
                    "type": "string",
                    "enum": [
                        "counter",
                        "gauge"
                    ],
                    "example": "gauge"
                },
                "type": {
                    "description": "параметр, принимающий значение gauge или counter",
                    "type": "string",
                    "example": "Alloc"
                },
                "value": {
                    "description": "значение метрики в случае передачи gauge",
                    "type": "number",
                    "example": 6649272
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    },
...
}